#13771 - Add Epipulse export functionality for Measles disease
